feat: create oracle manager as interface
